Refactor test helper related to logical driver change (#739)
* Refactoring to shared test area


* Add license header


---------
