+++ test/shared/com/xilinx/rapidwright/design/TestDesignHelper.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2023, Advanced Micro Devices, Inc.
+ * All rights reserved.
+ *
+ * Author: Chris Lavin, AMD Research and Advanced Development.
+ *
+ * This file is part of RapidWright.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.xilinx.rapidwright.design;
+
+import com.xilinx.rapidwright.device.Device;
+
+public class TestDesignHelper {
+
+    public static Net createTestNet(Design design, String netName, String[] pips) {
+        Net net = design.createNet(netName);
+        TestDesignHelper.addPIPs(net, pips);
+        return net;
+    }
+
+    public static void addPIPs(Net net, String[] pips) {
+        Device device = net.getDesign().getDevice();
+        for (String pip : pips) {
+            net.addPIP(device.getPIP(pip));
+        }
+    }
+
+}
